Bet Implement
Player can now bet any amount and player's blackjack is now accounted for. Win for x1, blackjack for x1.5, double down for x2 of bet amount.

Needs implementation: Dealer's 21 needs to be accounted for. Need to make a secret card that is drawn and caculated, but not added/displayed to the dealer's hand.